Move pool data shapes and DiluteLimitWarning out of defect_system
Pure code motion: extract the pool data-shape, normalisation and
serialisation plumbing into a new py_sc_fermi.pool_data module, and
relocate DiluteLimitWarning to py_sc_fermi.warnings. Both are re-imported
into defect_system, so the public surface (defect_system.DiluteLimitWarning
and the pool type aliases) is unchanged. No behaviour change -- the test
suite passes unedited.
